backend/requirements.txt
flask
Flask-SQLAlchemypip
Flask-Bcrypt
PyJWT
email-validator
flask-jwt-extended